Move non-parser functionality into the util package.
