+++ jplag.frontend.emf-metamodel-dynamic/src/main/java/de/jplag/emf/dynamic/parser/DynamicEcoreParser.java
@@ -5,8 +5,8 @@ import org.eclipse.emf.ecore.EObject;
 import de.jplag.emf.MetamodelToken;
 import de.jplag.emf.dynamic.DynamicMetamodelToken;
 import de.jplag.emf.dynamic.Language;
-import de.jplag.emf.parser.AbstractMetamodelVisitor;
 import de.jplag.emf.parser.EcoreParser;
+import de.jplag.emf.util.AbstractMetamodelVisitor;
 
 /**
  * Parser for EMF metamodels based on dynamically created tokens.
