+++ jplag.frontend.emf-metamodel-dynamic/src/main/java/de/jplag/emf/dynamic/parser/DynamicMetamodelTokenGenerator.java
@@ -3,7 +3,7 @@ package de.jplag.emf.dynamic.parser;
 import org.eclipse.emf.ecore.EObject;
 
 import de.jplag.emf.dynamic.DynamicMetamodelTokenConstants;
-import de.jplag.emf.parser.AbstractMetamodelVisitor;
+import de.jplag.emf.util.AbstractMetamodelVisitor;
 
 /**
  * Visits a metamodel containment tree and extracts the relevant token based on a dynamically created token set.
